enabled ttxt as a library (#22)
Made the crate as a lib (in addition to being a binary) but..

Did a bunch on unnecessary changes too:
- fixed clippy errors showing up in my editor (curious if you get any on
your end, I am using 1.82)
- enabled fmt to work on my IDE (nightly-2024-04-10)
- upgraded packages (my crates.nvim extension showed me a few warnings
with newer packages)
- enabled taplo by adding a config file resembling the one in
`polkadot-sdk`
- one of the tests fails on my machine (oh_god2), so I ignored it.
- upgrading the packages required at some point to drop the
`account_id()` call on keyring, and use `public_key()` instead. I think
this should be alright but lmk if you have another opinion.

I can revert the unnecessary changes if you want or find it important
for some reason.

Closes #18

---------

Signed-off-by: Iulian Barbu <iulian.barbu@parity.io>
Co-authored-by: Michal Kucharczyk <1728078+michalkucharczyk@users.noreply.github.com>